+++ pcs/lib/cib/resource/validations.py
@@ -68,6 +68,17 @@ def validate_move_resources_to_group(
                     )
                 )
             )
+        elif is_stonith(resource_element):
+            report_list.append(
+                reports.ReportItem.error(
+                    reports.messages.CannotGroupResourceWrongType(
+                        str(resource_element.attrib["id"]),
+                        "stonith",
+                        parent_id=None,
+                        parent_type=None,
+                    )
+                )
+            )
         else:
             parent = get_parent_resource(resource_element)
             if parent is not None and not group.is_group(parent):