+++ pcs_test/tier1/cib_resource/test_clone_unclone.py
@@ -334,18 +334,8 @@ class Clone(
 
     def test_clone_id_is_stonith(self):
         self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
+        self.assert_pcs_fail_regardless_of_force(
             "resource clone fence-device".split(),
-            "Error: This command does not accept stonith resources.\n",
-        )
-        self.assert_resources_xml_in_cib(
-            fixture_resources_xml(FIXTURE_STONITH_FOR_CLONE)
-        )
-
-    def test_clone_id_is_stonith_forced(self):
-        self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
-            "resource clone fence-device --force".split(),
             "Error: This command does not accept stonith resources.\n",
         )
         self.assert_resources_xml_in_cib(
@@ -483,18 +473,8 @@ class Clone(
 
     def test_promotable_clone_id_is_stonith(self):
         self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
+        self.assert_pcs_fail_regardless_of_force(
             "resource promotable fence-device".split(),
-            "Error: This command does not accept stonith resources.\n",
-        )
-        self.assert_resources_xml_in_cib(
-            fixture_resources_xml(FIXTURE_STONITH_FOR_CLONE)
-        )
-
-    def test_promotable_clone_id_is_stonith_forced(self):
-        self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
-            "resource promotable fence-device --force".split(),
             "Error: This command does not accept stonith resources.\n",
         )
         self.assert_resources_xml_in_cib(