+++ pcs_test/tier1/test_status.py
@@ -5,18 +5,15 @@ from unittest import TestCase
 from pcs import settings
 
 from pcs_test.tools.assertions import AssertPcsMixin
-from pcs_test.tools.bin_mock import get_mock_settings
 from pcs_test.tools.misc import get_test_resource as rc
 from pcs_test.tools.misc import (
     get_tmp_file,
     is_minimum_pacemaker_version,
     is_pacemaker_21_without_20_compatibility,
     outdent,
-    write_data_to_tmpfile,
     write_file_to_tmpfile,
 )
 from pcs_test.tools.pcs_runner import PcsRunner
-from pcs_test.tools.xml import XmlManipulation
 
 PCMK_2_0_3_PLUS = is_minimum_pacemaker_version(2, 0, 3)
 
@@ -201,47 +198,6 @@ class StonithWarningTest(TestCase, AssertPcsMixin):
 
                     WARNINGS:
                     No stonith devices and stonith-enabled is not false
-
-                    Stack: unknown
-                    Current DC: NONE
-                """
-                ),
-            )
-
-    def test_no_stonith_warning_when_stonith_in_group(self):
-        xml_manip = XmlManipulation.from_file(self.temp_cib.name)
-        xml_manip.append_to_first_tag_name(
-            "resources",
-            """
-            <group id="G">
-                <primitive class="stonith" id="S" type="fence_xvm">
-                    <operations>
-                        <op id="S-monitor-interval-60s" interval="60s"
-                            name="monitor"/>
-                    </operations>
-                </primitive>
-            </group>
-            """,
-        )
-        write_data_to_tmpfile(str(xml_manip), self.temp_cib)
-        self.pcs_runner.corosync_conf_opt = self.corosync_conf
-        self.pcs_runner.mock_settings = get_mock_settings("crm_resource_exec")
-        if PCMK_2_0_3_PLUS:
-            self.assert_pcs_success(
-                ["status"],
-                stdout_start=dedent(
-                    """\
-                    Cluster name: test99
-                    Cluster Summary:
-                """
-                ),
-            )
-        else:
-            self.assert_pcs_success(
-                ["status"],
-                stdout_start=dedent(
-                    """\
-                    Cluster name: test99
 
                     Stack: unknown
                     Current DC: NONE